+++ src/display_manager.h
@@ -57,6 +57,7 @@ int display_manager_active_display_count(void);
 uint32_t *display_manager_active_display_list(int *count);
 uint32_t display_manager_focus_display_with_point(uint32_t did, CGPoint point, bool update_cursor_position);
 void display_manager_focus_display(uint32_t did, uint64_t sid);
+enum space_op_error display_manager_focus_space(uint32_t did, uint64_t sid);
 bool display_manager_begin(struct display_manager *dm);
 
 #endif